chore: refactor Playwright E2E fixtures into two-layer architecture
Resolves #141

- Add e2e/fixtures/index.ts: two-layer fixture system (browserTest + tauriTest)
  exported via Playwright test.extend(). Specs no longer call setupTauriMock()
  directly — the Tauri mock, workspace injection, and app navigation are all
  handled by the tauriTest fixture.
- Per-test isolated workspace directories via fs.mkdtempSync, cleaned up in
  fixture teardown, exposed as the `workspaceDir` fixture.
- waitForAppReady fixture probes the React root mounting — no bare
  page.waitForTimeout() for app-ready signals in the migrated specs.
- Delete e2e/tests/editor.spec.ts (214-line monolith).
- Replace with 5 focused spec files under e2e/tests/editor/:
  - fixture-contract.spec.ts — structural tests for the new fixture API
  - editor-content.spec.ts — opening files and content rendering
  - typing.spec.ts — typing text in the editor
  - slash-commands.spec.ts — slash command menu and heading insertion
  - find-bar.spec.ts — find bar open/search/close
- Shared openFileInEditor helper exported from fixtures/index.ts to avoid
  duplication across spec files.
- Implementation is clean-room — no AGPL code copied or derived.
